update types references and split up types files
