+++ src/types/searchResult.ts
@@ -1,0 +1,15 @@
+export interface SearchResult {
+  Title: string;
+  Year: string;
+  Type: string;
+  Plot?: string;
+  imdbID: string;
+}
+
+export interface FormattedSearchResult {
+  Title: string;
+  Year: string;
+  Type: string;
+  Plot?: string;
+  'IMDb ID': string;
+}
